% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateCluster.R
\name{CalculateAUC}
\alias{CalculateAUC}
\title{Evaluate cluster effectiveness by AUC}
\usage{
CalculateAUC(
  data,
  cores = 1,
  min_pct = 0.05,
  logfc_threshold = 0.1,
  auc_cutoff = 0.6
)
}
\arguments{
\item{data}{A SeuratObject containing multiple resolution clusters.}

\item{cores}{Windows default 1. other system is free.}

\item{min_pct}{The parameter of FindAllMarkers, default 0.05. It will affect the result of AUC.Not recommended to change}

\item{logfc_threshold}{The parameter of FindAllMarkers, default 0.1. It will affect the result of AUC.Not recommended to change}

\item{auc_cutoff}{The threshold of AUC score, default 0.6. It will affect the maximum resolution.}
}
\value{
Return a list include the result of all methods
}
\description{
Calculate the AUC of Markers to help choose the max resolution.
}
\examples{
\dontrun{
CalculateAUC(data, cores=40)
CalculateAUC(data, cores=8, min_pct = 0.05, logfc_threshold = 0.2)
CalculateAUC(data, cores=1, min_pct = 0.1, logfc_threshold = 0.1)
}
}
