% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateCluster.R
\name{Choose_res}
\alias{Choose_res}
\title{Choose the resolution}
\usage{
Choose_res(result_eva, marker_counts, rogue_threshold = 0.85)
}
\arguments{
\item{result_eva}{the result of EvaluateCluster}

\item{marker_counts}{the result of CalculateAUC}

\item{rogue_threshold}{The threshold of rogue score, default 0.8. It will affect the minimum resolution}
}
\value{
Return a list include the result of all methods
}
\description{
Choose the best resolution based on the result of other function.
}
\examples{
\dontrun{
Choose_res(result_eva, marker_counts)
Choose_res(result_eva, marker_counts, rogue_threshold = 0.8)
Choose_res(result_eva, marker_counts, rogue_threshold = 0.9)
}
}
