% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateCluster.R
\name{EvaluateCluster}
\alias{EvaluateCluster}
\title{Evaluate cluster effectiveness}
\usage{
EvaluateCluster(
  data,
  cores = 1,
  rogue = FALSE,
  lisi = FALSE,
  mrtree = FALSE,
  CH_index = TRUE,
  plot = TRUE
)
}
\arguments{
\item{data}{A SeuratObject containing multiple resolution clusters}

\item{cores}{Windows default 1. other system is free}

\item{rogue}{First method, default false, recommend Windows to False}

\item{lisi}{Second method, default false}

\item{mrtree}{Third method, default false, recommend Windows to False}

\item{CH_index}{Fourth method, default true}

\item{plot}{Determine whether the function gives the plot of result}
}
\value{
Return a list include the result of 4 methods
}
\description{
Help choose the best clustering resolution in Seurat.
}
\examples{
\dontrun{
EvaluateClusters(data,cores=20)
EvaluateClusters(data,cores=8,rogue=FALSE,plot=FALSE)
EvaluateClusters(data,rogue=FALSE,mrtree=FALSE,lisi=FALSE,cores=1)
}
}
