% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateCluster.R
\name{EvaluateCluster}
\alias{EvaluateCluster}
\title{Evaluate cluster effectiveness}
\usage{
EvaluateCluster(data, cores = 1, auc_cutoff = 0.6, rogue_threshold = 0.8)
}
\arguments{
\item{data}{A SeuratObject containing multiple resolution clusters}

\item{cores}{Windows default 1. other system is free}

\item{auc_cutoff}{The threshold of AUC score, default 0.6. It will affect the maximum resolution}

\item{rogue_threshold}{The threshold of rogue score, default 0.8. It will affect the minimum resolution}
}
\value{
Return a list include the result of all methods
}
\description{
Help choose the best clustering resolution in Seurat.
}
\examples{
\dontrun{
EvaluateClusters(data, cores=40)
EvaluateClusters(data, cores=8, rogue_threshold=0.8, auc_cutoff=0.6)
EvaluateClusters(data, cores=1, rogue_threshold=0.85, auc_cutoff=0.7)
}
}
